calculate invoice maybe works i hope soi very very much

+++ src/main/java/DBook/project/app/offers/Invoice.java
@@ -38,9 +38,9 @@ public class Invoice implements Transactionable {
 
     }
 
-    public Integer calculateInvoice(Transaction txt){
+    public Float calculateInvoice(Transaction txt){
         Money result = new Money();
-        this.books.forEach((k, v) ->result.add(Integer.parseInt(v.getFromDB(txt).list().get(1).toString())));
+        this.books.forEach((k, v) ->result.add(Float.parseFloat(v.getFromDB(txt).list().get(1).toString())));
         return result.getValue();
     }
